backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
